slightly better handling for sqlite case.

+++ tests/conftest.py
@@ -14,7 +14,11 @@ from syrupy.extensions.single_file import SingleFileSnapshotExtension
 from lineapy import save
 from lineapy.data.types import SessionType
 from lineapy.db.db import RelationalLineaDB
-from lineapy.db.utils import MEMORY_DB_URL, resolve_default_db_path
+from lineapy.db.utils import (
+    DB_URL_ENV_VARIABLE,
+    MEMORY_DB_URL,
+    resolve_default_db_path,
+)
 from lineapy.execution.executor import Executor
 from lineapy.instrumentation.tracer import Tracer
 from lineapy.logging import configure_logging
@@ -248,14 +252,19 @@ def chdir_test_file():
     os.chdir(current_working_dir)
 
 
-@pytest.fixture(autouse=False)
+@pytest.fixture(autouse=True)
 def remove_db():
     """
     Remove db before all tests
     """
-    p = resolve_default_db_path()
-    if p.exists():
-        p.unlink()
+    # doing this because db cleanup is only needed for sqlite
+    db_url = (
+        os.environ.get(DB_URL_ENV_VARIABLE, MEMORY_DB_URL) or MEMORY_DB_URL
+    )
+    if db_url.startswith("sqlite"):
+        p = resolve_default_db_path()
+        if p.exists():
+            p.unlink()
 
 
 @pytest.fixture
